+++ bsp/stm32/stm32l475-atk-pandora/board/Kconfig
@@ -106,14 +106,34 @@ menu "Onboard Peripheral Drivers"
             default y
     endif
 
-    config BSP_USING_SDCARD
-        bool "Enable SDCARD (spi1)"
-        select BSP_USING_SPI
-        select BSP_USING_SPI1
-        select RT_USING_SPI_MSD
+
+
+    menuconfig BSP_USING_FS
+        bool "Enable File System"
         select RT_USING_DFS
-        select RT_USING_DFS_ELMFAT
+        select RT_USING_DFS_ROMFS
         default n
+
+        if BSP_USING_FS
+            config BSP_USING_SDCARD_FATFS
+                bool "Enable SDCARD (FATFS)"
+                select BSP_USING_SPI
+                select BSP_USING_SPI1
+                select RT_USING_SPI_MSD
+                select RT_USING_DFS_ELMFAT
+                default n
+
+            config BSP_USING_SPI_FLASH_LITTLEFS
+                bool "Enable SPI-FLASH (LittleFS)"
+                select RT_USING_MTD_NOR
+                select BSP_USING_QSPI_FLASH
+                select RT_USING_FAL
+                select FAL_USING_AUTO_INIT
+                select FAL_PART_HAS_TABLE_CFG
+                select PKG_USING_LITTLEFS
+                default n
+        endif
+
 
     config BSP_USING_ICM20608
         bool "Enable ICM20608 (i2c3)"
