+++ bsp/stm32/stm32l475-atk-pandora/board/SConscript
@@ -12,14 +12,20 @@ board.c
 CubeMX_Config/Src/stm32l4xx_hal_msp.c
 ''')
 
+path =  [cwd]
+path += [cwd + '/CubeMX_Config/Inc']
+
 if GetDepend(['BSP_USING_KEY']):
     src += Glob('ports/drv_key.c')
 
 if GetDepend(['BSP_USING_QSPI_FLASH']):
     src += Glob('ports/drv_qspi_flash.c')
     
-if GetDepend(['BSP_USING_SDCARD']):
-    src += Glob('ports/drv_sdcard.c')
+if GetDepend(['BSP_USING_FS']):
+    src += Glob('ports/drv_filesystem.c')
+    if GetDepend(['BSP_USING_SPI_FLASH_LITTLEFS']):
+        src += Glob('ports/fal/fal_spi_flash_sfud_port.c')
+        path += [cwd + '/ports/fal']
 
 if GetDepend(['BSP_USING_ICM20608']) or GetDepend(['BSP_USING_AHT10']):
     src += Glob('ports/sensor_port.c')
@@ -34,8 +40,7 @@ if GetDepend(['BSP_USING_AUDIO_RECORD']):
 if GetDepend(['BSP_USING_STM32_SDIO']):
     src += Glob('ports/drv_sdio_adapter.c')    
 
-path =  [cwd]
-path += [cwd + '/CubeMX_Config/Inc']
+
 
 if GetDepend(['BSP_USING_AUDIO']):
     path += [cwd + '/ports/audio']
